requirements.txt
alabaster==0.7.12
appdirs==1.4.3
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
dataclasses==0.7
docopt==0.6.1
docopts==0.6.1
docutils==0.16
emoji==0.5.4
idna==2.8
imagesize==1.2.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.1
paramiko==2.7.1
pathspec==0.7.0
pbr==5.4.4
pycparser==2.19
Pygments==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.6
pytz==2019.3
regex==2020.1.8
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scp==0.13.2
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.1
sphinxcontrib-apidoc==0.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
yummy-sphinx-theme==0.1.1
